+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## Unreleased
 
+## 0.2.1
+
+- Fix: Always emit permissions for main file/directory [#9](https://github.com/schneems/path_facts/pull/9). Previously we skipped emitting permissions when RWX were all true, but it looked off in some scenarios:
+
+
+```
+                  ├── `link_to_dir` directory []
+```
+
+Fixed to always output the permissions for the target file/directory:
+
+```
+                  ├── `link_to_dir` directory [✅ read, ✅ write, ✅ execute]
+```
+
 ## 0.2.0
 
 - Add: Minimum supported Rust version (MSRV) constraint as 1.79, which is when [`std::path::absolute`](https://doc.rust-lang.org/std/path/fn.absolute.html) was stabilized [#8](https://github.com/schneems/path_facts/issues/8).